+++ packages/tiny-canvas/src/index.js
@@ -3,4 +3,6 @@ import './style.css';
 export { default as Canvas } from './Canvas';
 export { default as Block } from './Block';
 export { default as Frame } from './Frame';
+export { default as Note } from './Note';
+export { default as Mark } from './Mark';
 export { useResetCanvas } from './useResetCanvas';